@@ -5,7 +5,7 @@ author: Akshath Jain <akshath@akshathjain.com>
 homepage: https://github.com/akshathjain/grouped_buttons
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=3.2.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -16,3 +16,4 @@ dev_dependencies:
     sdk: flutter
 
 flutter:
+  uses-material-design: true
