@@ -22,26 +22,26 @@ class RadioButtonGroup extends StatefulWidget {
   /// Specifies which buttons should be disabled.
   /// If this is non-null, no buttons will be disabled.
   /// The strings passed to this must match the labels.
-  final List<String> disabled;
+  final List<String>? disabled;
 
   /// Called when the value of the RadioButtonGroup changes.
-  final void Function(String label, int index) onChange;
+  final void Function(String label, int index)? onChange;
 
   /// Called when the user makes a selection.
-  final void Function(String selected) onSelected;
+  final void Function(String selected)? onSelected;
 
   /// The style to use for the labels.
   final TextStyle labelStyle;
 
   /// Specifies the orientation to display elements.
-  final GroupedButtonsOrientation orientation;
+  final GroupedButtonsOrientation? orientation;
 
   /// Called when needed to build a RadioButtonGroup element.
-  final Widget Function(Radio radioButton, Text label, int index) itemBuilder;
+  final Widget Function(Radio radioButton, Text label, int index)? itemBuilder;
 
   //RADIO BUTTON FIELDS
   /// The color to use when a Radio button is checked.
-  final Color activeColor;
+  final Color? activeColor;
 
   //SPACING STUFF
   /// Empty space in which to inset the RadioButtonGroup.
@@ -51,9 +51,9 @@ class RadioButtonGroup extends StatefulWidget {
   final EdgeInsetsGeometry margin;
 
   RadioButtonGroup({
-    Key key,
-    @required this.labels,
-    this.picked,
+    super.key,
+    required this.labels,
+    required this.picked,
     this.disabled,
     this.onChange,
     this.onSelected,
@@ -63,91 +63,90 @@ class RadioButtonGroup extends StatefulWidget {
     this.itemBuilder,
     this.padding = const EdgeInsets.all(0.0),
     this.margin = const EdgeInsets.all(0.0),
-  }) : super (key: key);
+  }) : super();
 
   @override
   _RadioButtonGroupState createState() => _RadioButtonGroupState();
 }
 
 class _RadioButtonGroupState extends State<RadioButtonGroup> {
-  String _selected;
+  late String _selected;
 
   @override
-  void initState(){
+  void initState() {
     super.initState();
 
     //set the selected to the picked (if not null)
-    _selected = widget.picked ?? "";
-
+    _selected = widget.picked;
   }
-
 
   @override
   Widget build(BuildContext context) {
-
-     //set the selected to the picked (if not null)
-    _selected = widget.picked ?? _selected;
-
+    //set the selected to the picked (if not null)
+    _selected = widget.picked;
 
     List<Widget> content = [];
-    for(int i = 0; i < widget.labels.length; i++){
-
+    for (int i = 0; i < widget.labels.length; i++) {
       Radio rb = Radio(
-                  activeColor: widget.activeColor ?? Theme.of(context).toggleableActiveColor,
-                  groupValue: widget.labels.indexOf(_selected),
-                  value: i,
+        activeColor:
+            widget.activeColor ?? Theme.of(context).colorScheme.primary,
+        groupValue: widget.labels.indexOf(_selected),
+        value: i,
 
-                  //just changed the selected filter to current selection
-                  //since these are radio buttons, and you can only pick
-                  //one at a time
-                  onChanged: (widget.disabled != null && widget.disabled.contains(widget.labels.elementAt(i))) ? null :
-                    (var index) => setState((){
-                      _selected = widget.labels.elementAt(i);
+        //just changed the selected filter to current selection
+        //since these are radio buttons, and you can only pick
+        //one at a time
+        onChanged: (widget.disabled != null &&
+                widget.disabled!.contains(widget.labels.elementAt(i)))
+            ? null
+            : (var index) => setState(() {
+                  _selected = widget.labels.elementAt(i);
 
-                      if(widget.onChange != null) widget.onChange(widget.labels.elementAt(i), i);
-                      if(widget.onSelected != null) widget.onSelected(widget.labels.elementAt(i));
-                    }),
-                );
-
-      Text t = Text(
-        widget.labels.elementAt(i),
-        style: (widget.disabled != null && widget.disabled.contains(widget.labels.elementAt(i))) ?
-                  widget.labelStyle.apply(color: Theme.of(context).disabledColor) :
-                  widget.labelStyle
+                  if (widget.onChange != null)
+                    widget.onChange!(widget.labels.elementAt(i), i);
+                  if (widget.onSelected != null)
+                    widget.onSelected!(widget.labels.elementAt(i));
+                }),
       );
 
+      Text t = Text(widget.labels.elementAt(i),
+          style: (widget.disabled != null &&
+                  widget.disabled!.contains(widget.labels.elementAt(i)))
+              ? widget.labelStyle.apply(color: Theme.of(context).disabledColor)
+              : widget.labelStyle);
+
       //use user defined method to build
-      if(widget.itemBuilder != null)
-        content.add(widget.itemBuilder(rb, t, i));
-      else{ //otherwise, use predefined method of building
+      if (widget.itemBuilder != null)
+        content.add(widget.itemBuilder!(rb, t, i));
+      else {
+        //otherwise, use predefined method of building
 
         //vertical orientation means Column with Row inside
-        if(widget.orientation == GroupedButtonsOrientation.VERTICAL){
-
+        if (widget.orientation == GroupedButtonsOrientation.VERTICAL) {
           content.add(Row(children: <Widget>[
             SizedBox(width: 12.0),
             rb,
             SizedBox(width: 12.0),
             t,
           ]));
-
-        }else{ //horizontal orientation means Row with Column inside
+        } else {
+          //horizontal orientation means Row with Column inside
 
           content.add(Column(children: <Widget>[
             rb,
             SizedBox(width: 12.0),
             t,
           ]));
-
         }
       }
-
     }
 
     return Container(
       padding: widget.padding,
       margin: widget.margin,
-      child: widget.orientation == GroupedButtonsOrientation.VERTICAL ? Column(children: content) : Row(children: content),
+      child: widget.orientation == GroupedButtonsOrientation.VERTICAL
+          ? Column(children: content)
+          : Row(children: content),
     );
   }
 }
